chnaged to mha gpu benchmark and dp scheduler

+++ mha_benchmark.py
@@ -10,7 +10,7 @@ def MHA_CPU(embed_dim, batch_size, seq_len, num_heads):
 
 def MHA_GPU(embed_dim, batch_size, seq_len, num_heads):
     mha = torch.nn.MultiheadAttention(embed_dim=embed_dim, num_heads=num_heads, batch_first=True).cuda()
-    x = torch.rand(batch_size, seq_len, embed_dim).cuda()
+    x = torch.rand(batch_size, seq_len, embed_dim).cuda().half()
 
     for _ in range(3):
         mha(x, x, x)
